chore: add dark styles for skeleton, add transition for detailed character page and renamed folder
