+++ src/components/ui/skeleton.tsx
@@ -3,7 +3,9 @@ interface Props {
 }
 
 const Skeleton: React.FC<Props> = ({ className }) => (
-  <div className={`animate-pulse bg-gray-200 ${className}`} />
+  <div
+    className={`animate-pulse bg-gray-200 dark:bg-neutral-800 ${className}`}
+  />
 );
 
 export default Skeleton;
