+++ src/main.tsx
@@ -7,7 +7,7 @@ import ErrorBoundary from "./components/errorBoundary";
 import HomePage from "./pages/homePage/homePage";
 import NotFountPage from "./pages/notFountPage";
 import Loader from "./components/ui/loader";
-import DetailedCharacter from "./pages/characterPage/detailedCharacter";
+import DetailedCharacter from "./pages/detailedCharacter/detailedCharacter";
 import ThemeProvider from "./components/themeProvider";
 
 const router = createBrowserRouter(
