+++ src/pages/detailedCharacter/useCharacterPage.tsx
